requirements/test.txt
# Test dependencies go here.
-r base.txt



coverage==4.4.1
flake8==3.5.0 # pyup: != 2.6.0
django-test-plus==1.0.20
factory-boy==2.9.2
django-coverage-plugin==1.5.0

# pytest
pytest-django==3.1.2
pytest-sugar==0.9.0
